basic express.js
Use of the get() method of the app object to handle routing.
A. argument is the path that is requested by the client.
B. The second argument is a callback function that gets called as a result of this request.  Our function simply responds with the status, which will be sent with the header,  and the HTML stored in our html variable, which will be sent as the response body.